fix debounce dependancy array
